+++ build.sh
@@ -9,11 +9,11 @@ else
     echo "ghc not found, not building $PROG.exe"
 fi
 
-if command -v hastec >/dev/null 2>&1; then
-    echo "building $PROG.js with hastec"
-    (cd src && hastec --make HasteMain.hs -o ../demo/$PROG.js) || exit 1
-else
-    echo "hastec not found, not building $PROG.js"
-fi
+#if command -v hastec >/dev/null 2>&1; then
+#    echo "building $PROG.js with hastec"
+#    (cd src && hastec --make HasteMain.hs -o ../demo/$PROG.js) || exit 1
+#else
+#    echo "hastec not found, not building $PROG.js"
+#fi
 
 ./build-packages.sh